// src/src/core/constants/enums.ts
export enum Reg {
  Pending = 0,
  Approved = 1,
  Queried = 2,
  Initiated = 3,
  Revalidated = 4,
  Assigned = 5,
  InspectionQueried = 6,
}

export enum Print {
  pending = 0,
  complete = 1,
}

export enum ActiveInactiveStatus {
  Active = 1,
  Inactive = 0,
}

export enum Status {
  Active = 1,
  Inactive = 0,
  Probation = 2,
  Suspended = 3,
}
// ...
export enum InspectionStatus {
  Cleared = 'cleared',
  Queried = 'queried',
  Assigned = 'assigned',
  Approved = 'approved',
  Missed = 'missed',
  ReInspection = 're-inspection',
}